Integrate linear CART and CSV export into parameter optimization (#57)
* Adjust python learning procedure to report hyperparameter-tuning results

* Close standard error stream

* Ignore empty messages

* Add additional options for python learner

* Log error for MLR

* Incorporate feedback regarding README

* Integrate linear decission tree regressor

* Fix data types for LinearCART

* Throw errors when errors appear while fitting

* Change type of max_features in RandomForest

+++ SPLConqueror/CommandLine/Commands.cs
@@ -194,7 +194,7 @@ namespace CommandLine
             if (components.Length > 1)
                 task = components[1];
 
-            string[] taskAsParameter = task.Split(new Char[] { ' ' });
+            string[] taskAsParameter = task.Trim().Split(new Char[] { ' ' });
 
             GlobalState.logInfo.logLine(COMMAND + line);
 
@@ -753,6 +753,10 @@ namespace CommandLine
                         // SVR, DecisionTreeRegression, RandomForestRegressor, BaggingSVR, KNeighborsRegressor, KERNELRIDGE, DecisionTreeRegressor
                         if (ProcessWrapper.LearningSettings.isLearningStrategy(taskAsParameter[0]))
                         {
+                            for (int i = 0; i < taskAsParameter.Length; i++)
+                            {
+                                taskAsParameter[i] = taskAsParameter[i].Replace("=", ":");
+                            }
                             handlePythonTask(true, configurationsLearning, taskAsParameter);
                         }
                         else
@@ -1071,6 +1075,19 @@ namespace CommandLine
                 printer.print(GlobalState.allMeasurements.Configurations, new List<NFProperty>());
                 printNFPsToFile(configurationsLearning, nfpLearnFile);
                 printNFPsToFile(GlobalState.allMeasurements.Configurations, nfpValFile);
+                bool performWithOptimal = true;
+                if (taskAsParameter.Contains("performWithOptimal:false"))
+                {
+                    taskAsParameter = taskAsParameter.Except(new []{"performWithOptimal:false"}).ToArray();
+                    performWithOptimal = false;
+                }
+
+                bool writeFinalResults = true;
+                if (taskAsParameter.Contains("writeFinalPredictionResults:false"))
+                {
+                    taskAsParameter = taskAsParameter.Except(new []{"writeFinalPredictionResults:false"}).ToArray();
+                    writeFinalResults = false;
+                }
                 PythonWrapper pyInterpreter = new PythonWrapper(Path.Combine(getLocationPythonScript(), 
                     PythonWrapper.COMMUNICATION_SCRIPT), taskAsParameter);
                 GlobalState.logInfo.logLine("Starting Prediction");
@@ -1082,15 +1099,18 @@ namespace CommandLine
                     string path = targetPath.Substring(0, (targetPath.Length
                         - (((targetPath.Split(Path.DirectorySeparatorChar)).Last()).Length)));
                     pyResult = pyInterpreter.getOptimizationResult(GlobalState.allMeasurements.Configurations, path);
-                    GlobalState.logInfo.logLine("Optimal parameters " + pyResult.Replace(",", ""));
-                    File.Delete(configsLearnFile);
-                    File.Delete(configsValFile);
-                    File.Delete(nfpLearnFile);
-                    File.Delete(nfpValFile);
-                    var optimalParameters = pyResult.Replace(",", "").Split(new char[] { ';' },
-                        StringSplitOptions.RemoveEmptyEntries).ToList();
-                    optimalParameters.Insert(0, taskAsParameter[0]);
-                    handlePythonTask(false, configurationsLearning, optimalParameters.ToArray());
+                    if (performWithOptimal)
+                    {
+                        GlobalState.logInfo.logLine("Optimal parameters " + pyResult.Replace(",", ""));
+                        File.Delete(configsLearnFile);
+                        File.Delete(configsValFile);
+                        File.Delete(nfpLearnFile);
+                        File.Delete(nfpValFile);
+                        var optimalParameters = pyResult.Replace(",", "").Split(new char[] { ';' },
+                            StringSplitOptions.RemoveEmptyEntries).ToList();
+                        optimalParameters.Insert(0, taskAsParameter[0]);
+                        handlePythonTask(false, configurationsLearning, optimalParameters.ToArray());   
+                    }
                 }
                 else
                 {
@@ -1104,14 +1124,18 @@ namespace CommandLine
 
                     pyInterpreter.setupApplication(configsLearnFile, nfpLearnFile, configsValFile, nfpValFile,
                         PythonWrapper.START_LEARN, GlobalState.varModel, treePath);
-                    PythonPredictionWriter csvWriter = new PythonPredictionWriter(targetPath, taskAsParameter,
-                        GlobalState.varModel.Name + "_" + samplingIdentifier);
+                    PythonPredictionWriter csvWriter = null;
+                    if (writeFinalResults)
+                    {
+                        csvWriter = new PythonPredictionWriter(targetPath, taskAsParameter,
+                            GlobalState.varModel.Name + "_" + samplingIdentifier);
+                    }
                     List<Configuration> predictedByPython;
                     double error = pyInterpreter.getLearningResult(GlobalState.allMeasurements.Configurations, csvWriter, out predictedByPython);
+                    GlobalState.logInfo.logLine("Elapsed learning time(seconds): " + pyInterpreter.getTimeToLearning());
 
                     if (File.Exists(treePath))
                     {
-                        GlobalState.logInfo.logLine("Elapsed learning time(seconds): " + pyInterpreter.getTimeToLearning());
                         while (pyInterpreter.isRunning())
                             System.Threading.Thread.Sleep(100);
                         StreamReader reader = new StreamReader(treePath);
@@ -1138,12 +1162,22 @@ namespace CommandLine
                         GlobalState.allMeasurements.Configurations = tmp;
                     }
 
-                    GlobalState.logInfo.logLine("Prediction finished, results written in " + csvWriter.getPath());
+                    pyInterpreter.finish();
+
+                    if (writeFinalResults)
+                    {
+                        GlobalState.logInfo.logLine("Prediction finished, results written in " + csvWriter.getPath());
+                        csvWriter.close();
+                    }
+                    else
+                    {
+                        GlobalState.logInfo.logLine("Prediction finished.");
+                    }
+                    
                     if (!Double.IsNaN(error))
                     {
-                        GlobalState.logInfo.logLine("Error rate: " + error);
+                        GlobalState.logInfo.logLine("Prediction Error: " + error + " %");
                     }
-                    csvWriter.close();
                 }
             }
             finally
